Update test script to allow standard arguments
Setup script to allow optional arguments to standard lint script, such
as a file or directory, or --fix flag.

Update script to use npx to run standard script rather than npm,
which does not handle both the fix flag and the directories when passed
as additional arguments into `script/test`.